+++ src/containers/sidebar/index.tsx
@@ -1,18 +1,28 @@
 'use client';
-import { FC } from 'react';
+import { FC, useMemo } from 'react';
 
-import { useMonitor } from '@/hooks/monitors';
+import { useSearchParams } from 'next/navigation';
+
+import { useMonitor, useMonitors } from '@/hooks/monitors';
 
 import Loading from '@/components/loading';
-import MonitorsList from '@/components/monitors-list';
+import MonitorDisplay from '@/components/monitors-datasets-display';
 import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs';
+
 export const Sidebar: FC = () => {
+  const { data: monitors } = useMonitors();
+  const defaultMonitor = monitors?.[0];
+  const params = useSearchParams();
+  const monitor_id = useMemo<string>(
+    () => params.get('monitor_id') || defaultMonitor?.id,
+    [params, defaultMonitor]
+  );
   const { data, isLoading, isFetched, isError } = useMonitor({
-    monitor_id: 'm1',
+    monitor_id: monitor_id,
   });
   return (
     <aside className="absolute bottom-16 left-5 top-5 z-50 w-[30vw] overflow-y-auto bg-brand-600 p-7.5">
-      <Tabs defaultValue="account">
+      <Tabs defaultValue="datasets">
         <TabsList>
           <TabsTrigger value="datasets">Datasets</TabsTrigger>
           <TabsTrigger value="geostories">Geostories</TabsTrigger>
@@ -20,7 +30,7 @@ export const Sidebar: FC = () => {
         <TabsContent value="datasets">
           {' '}
           {isLoading && <Loading visible={isLoading} />}
-          {isFetched && !isError && <MonitorsList monitor={data} />}
+          {isFetched && !isError && <MonitorDisplay monitor={data} />}
         </TabsContent>
         <TabsContent value="geostories"> </TabsContent>
       </Tabs>