+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Unreleased
 - Satoshi Font, main navigation, and test for navigation functionality [OEMC-18](https://vizzuality.atlassian.net/browse/OEMC-18)
 - Map set up and layers display [OEMC-23](https://vizzuality.atlassian.net/jira/software/c/projects/OEMC/boards/95?selectedIssue=OEMC-23)
 - Map controls to move around the map (zoom in, out, fit bounds)[OEMC-44](https://vizzuality.atlassian.net/browse/OEMC-44?atlOrigin=eyJpIjoiMGYzNzk5OWQwNmExNDRkNjllOWE5NWMxNjc4MmIwMmQiLCJwIjoiaiJ9)
+- Toggle layers displayed on the map [OEMC-38](https://vizzuality.atlassian.net/browse/OEMC-58?atlOrigin=eyJpIjoiMTgyZDgyNzgzNzBhNGMxODljZThjYTk0YTQ3N2VkYzciLCJwIjoiaiJ9)
+- Datasets display on the map [OECM-62](https://vizzuality.atlassian.net/browse/OEMC-62?atlOrigin=eyJpIjoiZThhZDBmZTQyYTJiNDM1ZmFmZWI4MzZhNGNjYjkzMWMiLCJwIjoiaiJ9)
+
 
 
 ### Changed